Update Host interface yaml
Host.interface.yaml defines the mechanism to communicate with
the Host via a method. Since the data associated with OCCReset
command is the sensor id of the OCC, there is need now to
augment the signature of the API to take this data portion.

OCC error monitor application will call Execute method with the
sensor ID of the failing OCC.

Change-Id: I9f532f3d565448a1840c5ba61402c7cddea0224c
Signed-off-by: Vishwanatha Subbanna <vishwa@linux.vnet.ibm.com>