Add FSI device driver scan errors
Add errors for when the FSI device driver master and
hub scans fail.

This does not refer to the scan operation that scans
rings out of the chip.

The master scan fails has a callout to indicate the
master processor failed, while the hub scan does not
because currently it cannot be isolated at all.

Change-Id: I4a14af71c03924c000663ce271465127e15aad19
Signed-off-by: Matt Spinler <spinler@us.ibm.com>